[XLA:GPU] Fix record_ffi's number_of_shared_bytes override on 64-bit Darwin
size_t and uint64_t are distinct types there, so the override did not
override.

+++ xla/backends/gpu/runtime/record_ffi.cc
@@ -120,7 +120,7 @@ class FfiKernelArgsPacked : public se::KernelArgsPackedArrayBase {
     return argument_addresses_.size() + (shared_memory_bytes_ > 0);
   }
 
-  size_t number_of_shared_bytes() const override {
+  uint64_t number_of_shared_bytes() const override {
     return shared_memory_bytes_;
   }
